--- a/pkg/local_object_storage/metabase/db.go
+++ b/pkg/local_object_storage/metabase/db.go
@@ -11,6 +11,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/mr-tron/base58"
 	v2object "github.com/nspcc-dev/neofs-api-go/v2/object"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard/mode"
 	"github.com/nspcc-dev/neofs-node/pkg/util/logger"
@@ -111,6 +112,8 @@ func stringifyValue(key string, objVal []byte) string {
 	switch key {
 	default:
 		return string(objVal)
+	case v2object.FilterHeaderObjectID, v2object.FilterHeaderContainerID, v2object.FilterHeaderParent:
+		return base58.Encode(objVal)
 	case v2object.FilterHeaderPayloadHash, v2object.FilterHeaderHomomorphicHash:
 		return hex.EncodeToString(objVal)
 	case v2object.FilterHeaderCreationEpoch, v2object.FilterHeaderPayloadLength:
@@ -140,6 +143,9 @@ func destringifyValue(key, value string, prefix bool) ([]byte, bool, bool) {
 	switch key {
 	default:
 		return []byte(value), false, true
+	case v2object.FilterHeaderObjectID, v2object.FilterHeaderContainerID, v2object.FilterHeaderParent:
+		v, err := base58.Decode(value)
+		return v, false, err == nil
 	case v2object.FilterHeaderPayloadHash, v2object.FilterHeaderHomomorphicHash:
 		v, err := hex.DecodeString(value)
 		if err != nil {
@@ -254,6 +260,12 @@ func unknownMatcherBucket(_ *bbolt.Bucket, _ string, _ string, _ func([]byte, []
 // in boltDB. Useful for getting filter values from unique and list indexes.
 func bucketKeyHelper(hdr string, val string) []byte {
 	switch hdr {
+	case v2object.FilterHeaderParent:
+		v, err := base58.Decode(val)
+		if err != nil {
+			return nil
+		}
+		return v
 	case v2object.FilterHeaderPayloadHash:
 		v, err := hex.DecodeString(val)
 		if err != nil {