--- a/pkg/local_object_storage/metabase/movable.go
+++ b/pkg/local_object_storage/metabase/movable.go
@@ -52,13 +52,16 @@ func (db *DB) ToMoveIt(prm ToMoveItPrm) (res ToMoveItRes, err error) {
 	db.modeMtx.RLock()
 	defer db.modeMtx.RUnlock()
 
+	key := make([]byte, addressKeySize)
+	key = addressKey(prm.addr, key)
+
 	err = db.boltDB.Update(func(tx *bbolt.Tx) error {
 		toMoveIt, err := tx.CreateBucketIfNotExists(toMoveItBucketName)
 		if err != nil {
 			return err
 		}
 
-		return toMoveIt.Put(addressKey(prm.addr), zeroValue)
+		return toMoveIt.Put(key, zeroValue)
 	})
 
 	return
@@ -69,13 +72,16 @@ func (db *DB) DoNotMove(prm DoNotMovePrm) (res DoNotMoveRes, err error) {
 	db.modeMtx.RLock()
 	defer db.modeMtx.RUnlock()
 
+	key := make([]byte, addressKeySize)
+	key = addressKey(prm.addr, key)
+
 	err = db.boltDB.Update(func(tx *bbolt.Tx) error {
 		toMoveIt := tx.Bucket(toMoveItBucketName)
 		if toMoveIt == nil {
 			return nil
 		}
 
-		return toMoveIt.Delete(addressKey(prm.addr))
+		return toMoveIt.Delete(key)
 	})
 
 	return