--- a/pkg/local_object_storage/metabase/shard_id.go
+++ b/pkg/local_object_storage/metabase/shard_id.go
@@ -6,7 +6,7 @@ import (
 )
 
 var (
-	shardInfoBucket = []byte(invalidBase58String + "i")
+	shardInfoBucket = []byte{shardInfoPrefix}
 	shardIDKey      = []byte("id")
 )