--- a/pkg/local_object_storage/metabase/version.go
+++ b/pkg/local_object_storage/metabase/version.go
@@ -8,7 +8,7 @@ import (
 )
 
 // version contains current metabase version.
-const version = 1
+const version = 2
 
 var versionKey = []byte("version")