Close #1482 .

```
name                        old time/op    new time/op    delta
Select/string_equal-8         8.87µs ± 5%    2.92µs ± 7%  -67.05%  (p=0.000 n=10+10)
Select/string_not_equal-8     4.95ms ±11%    1.34ms ± 9%  -73.04%  (p=0.000 n=10+10)
Select/common_prefix-8        55.5µs ± 6%    15.5µs ± 9%  -72.08%  (p=0.000 n=9+10)
Select/unknown-8              3.32µs ± 9%    0.85µs ± 8%  -74.46%  (p=0.000 n=10+10)
Put/parallel-8                1.27ms ±12%    1.13ms ±19%  -11.07%  (p=0.009 n=10+10)
Put/sequential-8              4.55ms ±14%    4.32ms ±12%     ~     (p=0.190 n=10+10)
ListWithCursor/1_item-8       4.96µs ±16%    3.25µs ± 7%  -34.32%  (p=0.000 n=10+10)
ListWithCursor/10_items-8     15.4µs ±14%     6.5µs ± 8%  -57.47%  (p=0.000 n=10+10)
ListWithCursor/100_items-8     119µs ± 8%      37µs ± 4%  -69.08%  (p=0.000 n=9+8)
Get/1_objects-8               8.46µs ±15%    3.96µs ±17%  -53.15%  (p=0.000 n=9+10)
Get/10_objects-8              92.8µs ±12%    41.8µs ±12%  -54.98%  (p=0.000 n=10+9)
Get/100_objects-8              944µs ±12%     415µs ± 8%  -56.05%  (p=0.000 n=10+9)

name                        old alloc/op   new alloc/op   delta
Put/parallel-8                 133kB ±17%     135kB ±10%     ~     (p=0.631 n=10+10)
Put/sequential-8               174kB ± 4%     186kB ± 4%   +6.74%  (p=0.000 n=10+9)
ListWithCursor/1_item-8       1.91kB ± 0%    1.49kB ± 0%  -21.80%  (p=0.000 n=10+10)
ListWithCursor/10_items-8     5.50kB ± 0%    2.77kB ± 0%  -49.59%  (p=0.002 n=8+10)
ListWithCursor/100_items-8    41.5kB ± 0%    15.7kB ± 0%  -62.20%  (p=0.000 n=10+10)
Get/1_objects-8               2.83kB ± 0%    2.16kB ± 0%  -23.73%  (p=0.000 n=9+10)
Get/10_objects-8              29.6kB ± 2%    22.7kB ± 0%  -23.16%  (p=0.000 n=10+10)
Get/100_objects-8              295kB ± 1%     227kB ± 0%  -23.20%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
Put/parallel-8                   478 ± 1%       343 ± 2%  -28.18%  (p=0.000 n=10+10)
Put/sequential-8                 811 ± 1%       670 ± 1%  -17.37%  (p=0.000 n=10+10)
ListWithCursor/1_item-8         34.0 ± 0%      27.0 ± 0%  -20.59%  (p=0.000 n=10+10)
ListWithCursor/10_items-8        100 ± 0%        51 ± 0%  -49.00%  (p=0.000 n=10+10)
ListWithCursor/100_items-8       758 ± 0%       290 ± 0%  -61.74%  (p=0.000 n=10+10)
Get/1_objects-8                 48.0 ± 0%      36.0 ± 0%  -25.00%  (p=0.000 n=9+10)
Get/10_objects-8                 506 ± 1%       389 ± 0%  -23.09%  (p=0.000 n=10+10)
Get/100_objects-8              5.07k ± 1%     3.89k ± 0%  -23.24%  (p=0.000 n=10+10)
```