@fyrchik 2022-09-09
New benchmarks -with the current master:
```
name                        old time/op    new time/op    delta
Get/1_objects-8               11.9µs ± 7%     4.4µs ± 3%  -62.94%  (p=0.000 n=8+9)
Get/10_objects-8               120µs ± 7%      56µs ±19%  -53.61%  (p=0.000 n=10+10)
Get/100_objects-8             1.23ms ± 6%    0.57ms ± 7%  -53.58%  (p=0.000 n=10+10)
ListWithCursor/1_item-8       5.04µs ± 8%    4.17µs ± 9%  -17.31%  (p=0.000 n=10+10)
ListWithCursor/10_items-8     16.1µs ±11%     8.1µs ± 3%  -49.63%  (p=0.000 n=10+8)
ListWithCursor/100_items-8     117µs ± 6%      46µs ± 7%  -60.51%  (p=0.000 n=10+10)
Put/parallel-8                1.56ms ± 8%    1.37ms ±28%  -12.06%  (p=0.002 n=10+10)
Put/sequential-8              5.03ms ± 6%    4.72ms ±14%     ~     (p=0.068 n=8+10)
Select/string_equal-8         11.5µs ± 3%     3.7µs ±10%  -67.63%  (p=0.000 n=9+10)
Select/string_not_equal-8     7.52ms ± 1%    1.82ms ± 3%  -75.82%  (p=0.000 n=8+7)
Select/common_prefix-8        85.9µs ± 3%    20.7µs ± 7%  -75.91%  (p=0.000 n=9+10)
Select/unknown-8              3.34µs ± 6%    1.05µs ± 9%  -68.51%  (p=0.000 n=10+9)

name                        old alloc/op   new alloc/op   delta
Get/1_objects-8               3.32kB ± 0%    1.86kB ± 0%  -43.86%  (p=0.000 n=9+10)
Get/10_objects-8              34.9kB ± 1%    23.2kB ± 1%  -33.56%  (p=0.000 n=9+10)
Get/100_objects-8              351kB ± 0%     234kB ± 0%  -33.26%  (p=0.000 n=8+10)
ListWithCursor/1_item-8       1.91kB ± 0%    1.49kB ± 0%  -21.80%  (p=0.000 n=10+10)
ListWithCursor/10_items-8     5.50kB ± 0%    2.77kB ± 0%  -49.59%  (p=0.000 n=10+10)
ListWithCursor/100_items-8    41.5kB ± 0%    15.7kB ± 0%  -62.20%  (p=0.000 n=7+10)
Put/parallel-8                 118kB ± 6%     131kB ± 9%  +11.17%  (p=0.000 n=9+10)
Put/sequential-8               171kB ± 1%     184kB ± 2%   +7.83%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
Get/1_objects-8                 57.0 ± 0%      36.0 ± 0%  -36.84%  (p=0.000 n=9+10)
Get/10_objects-8                 606 ± 1%       408 ± 0%  -32.58%  (p=0.000 n=10+10)
Get/100_objects-8              6.10k ± 0%     4.09k ± 0%  -32.95%  (p=0.000 n=8+10)
ListWithCursor/1_item-8         34.0 ± 0%      27.0 ± 0%  -20.59%  (p=0.000 n=10+10)
ListWithCursor/10_items-8        100 ± 0%        51 ± 0%  -49.00%  (p=0.000 n=10+10)
ListWithCursor/100_items-8       758 ± 0%       290 ± 0%  -61.74%  (p=0.000 n=10+10)
Put/parallel-8                   490 ± 1%       335 ± 1%  -31.62%  (p=0.000 n=10+10)
Put/sequential-8                 833 ± 0%       672 ± 1%  -19.38%  (p=0.000 n=10+10)
```