@fyrchik 2022-09-15
Rebased on master